ensure fixed Discounts are applied to each item instead of once
You can customize the `.quantity()`, but it will default to the quantity
of the current `invoice item` in case of a fixed price. For example, if
you get €1,- discount for each chocolate bar, and you buy 5 chocolate
bars, you get a total discount of €5,- not €1,- that it used to be.