ensure `discount.quantity` is always a number for `fixed` discounts
It used to be `number | null`, but defaulting to `1` makes more sense.
This also means that we should be able to differentiate between the
default `1` and the explicitly set `1`.

For that reason, a `quantityType` is introduced, but 99% of the code
doesn't use that particular part and therefore `quantity` being a number
is the better solution.