use smarter parsing for tables
Currently we are splitting all DOM nodes deeply nested such that we can
split anywhere in the document. However, `<tr>` are considered rows that
don't wrap, so there is no need to deeply split the `<td>`'s as well.

This also means there is less work involved in expanding, and therefore
also less work for re-stringifying.