move all remaining items to the next page
If we don't do this, then it can happen that while rendering the system
thinks there is only 1 page, even though there should be more.

Sometimes we have code that says "if it is the last page render X".
These assumptions would now be wrong.

This commit fixes that because if there are more items, a new page
_will_ be created.