add `Promote to Invoice` button
Currently all the data is coming from code, there is no database so some
things aren't super easy to do right now.

A first step to creating an invoice from a quote could be this little
form. But instead of saving this to a database, you get the code snippet
instead that you can paste in your data file.