add loading indicator
Currently a hack to work around a UI "bug" where reading data from
localStorage is too late (after the initial render). Therefore, in this
case the sidebar will start "open", and then close depending on the data
in localStorage.

Ideally we set this information in a cookie or something and render it
on the server based on the cookie as well such that SSR is correct.