improve activity feed card
When scrolling, the activity items should touch the top/bottom of the
card. But when you are all the way at the top / bottom the padding
should be there.

This makes the UI nicer while scrolling because the little vertical line
indicators will be touching the top/bottom as expected.