Add milestones (#8)
* add events to `Account` and `Client`

* only render activity feed if we do have events

* always show the event icon in the activity feed

* add icons for relocation and rebranded events

* add optional timestamps to relocate and rebrand actions

* allow rendering additional data in activity feed

* fix small bug: set new formatter correctly

Before it was calling the function already because of this signature:

```js
setFormatter((previous) => {
   return /* ... */
})
```

* milestones approach 1

* milestones approach 2 (refactor)

* ensure `lazy-collections` is transpiled when running tests

* improve English

* refactor milestones system

* add private `InvoiceBuilder.emit`

* add private `QuoteBuilder.emit`

* add private `ReceiptBuilder.emit`

* add "fastest paid invoice" milestone

* disable autoComplete for command palette input

* bump to latest TypeScript

* add "fastest accepted quote" milestone

* make TypeScript happy

Necessary because we bumped TypeScript to 5.x

* export each milestone tracker

* drop useless check

We can't accept a quote if it was not sent, so this check will never be
`true`.

* add milestone tests

* use a simple `Map` to track the totalByInvoice

* fix incorrect logic

Otherwise invoices with the exact same amount could result in the same
milestone being achieved.

* improve responsive layout

* reference correct items

They should in theory be the same, but let's use the correct reference
at least.

* add events to `client`, `quote` and `invoice as well

* show the indicator for the last item OR the last non-milestone event

* use the currency of the client when looking at the client page

Otherwise the money amounts could be the incorrect ones compared to the
invoice itself.

* return first milestone

It could be that we reach multiple milestones using a single invoice. In
this case we want to make sure to only use the _largest_ milestone.

The deleted list was from hight to low, which means that using `pop`
took the smallest.

* add 500 and 100 as revenue milestones

* make `future` a boolean

It's optional when you want to generate an Event, but the output will be
a boolean (defaulted to `false`).

* add tags to events + update tests

* use snapshot tests for milestone expectations

* use `Europe/Brussels` as timezone for testing

* use tags instead of checking the type

* move TZ env to package.json 🤔

* simplify `account-milestone:xxx` to `milestone:xxx`

This will allow us to render the milestone in context of another entity.
E.g.: Most expensive invoice will be rendered differently on the
`account` page, compared to the `client` page.

* replace `embedded` with `filled`

* improve activity feed card

When scrolling, the activity items should touch the top/bottom of the
card. But when you are all the way at the top / bottom the padding
should be there.

This makes the UI nicer while scrolling because the little vertical line
indicators will be touching the top/bottom as expected.

* add "most expensive invoice" to the events of the client as well

* group related events together

* add view context to `ActivityFeed`

This will allow us to slightly change the messages of the events
depending where you are looking at them.

* set ViewContext in ActivityFeed for records

* add `best` bool to revenue milestone

This will be set to false for all revenue milestones in the past since
they are not the most expensive one.

* use 24/outline for templates eye icon

* render events slightly differently based on viewContext

* refactor resolving of milestones

This will allow us to easily generate events for multiple milestones.
Right now it will result in 0 or 1 milestone.

* improve wording based on viewContext

* update tests with new `best` property

* introduce `tombstone` field on `Event`

This is not ideal, but until we have a better data model for events this
will do.

* rethink events

- Don't store the events themselves in the Account / Client / Record
  entities.
- Store events in a shared location:
   - Should be easier in the future as well if we in the future store it
     in a DB.
- Dispatch all events using shared EventEmitter.
- Implement milestones based on these events, re-emit milestones events
  as a chain reaction.

* bump zod

* refactor events

* add `international-clients` milestone

* separate event names with `:`

E.g.: `quote:accepted` instead of `quote-accepted`

* format

* remove unused `on` from disposables

* add `anniversary` milestone

* ensure `isFutureEvent` is looking at the correct data