track the first invoice/quote
We don't want to create a milestone for the very first "fastest accepted
quote" or "fastest paid invoice" because otherwise the first invoice
will retrieve a bunch of milestones because it beats all of the
milestones by being the only one.

However, we were also not really tracking them, so the next paid
invoices was always the fastest even though the first one could have
been faster.

This fixes that, by tracking the data but not emitting a milestone event.