fix select/copy on money amounts in PDFs
The `font-sans` on the `<Money />` component results in non-copyable text
after generating a PDF and opening it in Chrome. Copying the value
results in just the `€` sign being copied.

This is likely a bug in either Chrome or in the font or a combination of
both. The default `font-sans` font is `Inter`.

We do have a `font-pdf` utility, which uses good ol' Arial, but doesn't
look as nice (even with `tabular-nums`).

Settled for the default `font-mono` instead which uses default fonts
from your OS. In the latest macOS, this resolves to `SF Mono`.